requirements.txt
# Core dependencies
ccxt>=4.0.0              # Unified API for crypto exchanges
pandas>=2.0.0            # Data manipulation
numpy>=1.24.0            # Numerical computing

# Visualization
streamlit>=1.28.0        # Interactive dashboard
plotly>=5.17.0          # Interactive charts
altair>=5.1.0           # Declarative visualization

# Data storage
sqlalchemy>=2.0.0       # Database ORM
pyyaml>=6.0             # Configuration files

# Utilities
python-dotenv>=1.0.0    # Environment variables
requests>=2.31.0        # HTTP requests
websockets>=12.0        # Real-time data streams

# Logging and monitoring
loguru>=0.7.0           # Better logging

# Development
pytest>=7.4.0           # Testing
black>=23.0.0           # Code formatting
flake8>=6.1.0           # Linting
